gui: remove "Type" column from Peers main window
as it is confusing to users and does not make sense
to have a Type column without a Direction column
immediately preceding it to indicate Inbound or Outbound.

Several attempts to add the Direction column, as
described, have seen opposition from reviewers.

Therefore, remove the Type column until such time as
there is rough consensus to add the Direction
Inbound/Outbound column with it.