qt: move third-party tx URL setting from Display to Wallet options tab
Out of all the settings in the Display tab, the 'Third-party
transaction URLs' is the odd one out. All other settings here
affect the way the GUI is displayed. This action affects what
context menu actions are available when wallet functionality
is enabled, and we have a transaction. As such, it is a better
fit for this to be in the Wallet tab.