qt: Remove dbus notification code
As of Qt 5.5, Qt will automatically use Freedesktop notifications when
available and relevant to the platform. There is no more need to roll
our own, so remove the dbus notification code.

This change also removes any mention of the QtDbus library from the
build system. We do not need to link against it explicitly anymore.
I have however not removed Qt dbus flags from the depends
system, because it needs to be available internally for this to work.

The feature to set a custom QIcon for notifications is also removed, as
it was only available for DBUS and besides, isn't used at all!