gui: Add Address Type Column to AddressBookPage
Add new address type column to the addresstablemodel, use the getOutputType
function from the wallet interface to display the correct address type in the
new column content. Update the address book page to set the new column resize mode
and to display the new column only when the receiving tab is enabled
so it must be hidden on the sending tab.

Update AddressBookFilterProxyModel::filterAcceptsRow so the filter works also
on the new addressType column content. Also the searLineEdit greyed text
reflects that the new field/ column addressType will be included in the search/
filter by but only when the receiving tab is enabled.

Add the new column to the export feature.