gui: Extend address book filter for nested filtering
Extend AddressBookFilterProxyModel to allow using nested filters to be applied
on top of it. If future needs arise for similar filters outside the address book page,
the code could be easily refactored and moved in a new subclass of QSortFilterProxyModel,
perhaps with limits on nested levels.

For safety and performance reasons, the code of the filter proxy model class declaration
(in addressbookpage.h) and its instance creation were updated by aligning it with
TransactionFilterProxy in overviewpage.h as this addresses situations of unexpected crashes,
such as segfault on closing the app, double free or corruption, or stack smashing detection.