gui: improve user dialog when signing multisig psbts
Prior to this commit, when a psbt requiring multiple signers
is signed via gui the dialog is vague/confusing. Whether the
signer successfully partially signs the psbt or not the same warning
is displayed "Could not sign any more inputs." After signing the tx,
the only way for the user to know if their action had any effect is
to inspect the psbt (unless it was the last signature that completed
the psbt, in which case there is a success dialog).

Now we indicate when a psbt has been partially signed, and show how many
partial signatures the psbt has in the transaction description.