Including exception what() in Runaway dialog box
If an exception happens, including e.what() in the text passed to the dialog
box, so that the user can see it.  This is in addition to any text from
getWarnings() that might already be present, separated by newlines if so.

This will make it easier for users to find what went wrong and get the help
they need, instead of having to tell the user to go back and dig through the
debug.log file.